requirements.txt
opencv-python
numpy
requests
Pillow
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
tqdm
nltk
pytesseract
google-generativeai>=0.3.0